refactor: change createJsonStorage to accept custom reviver/replacer
